refactor out of manager
