+++ lib/config/index.js
@@ -7,10 +7,25 @@ const envParser = require('./env');
 
 const { getPlatformApi } = require('../platform');
 const { resolveConfigPresets } = require('./presets');
+const { get, getLanguageList, getManagerList } = require('../manager');
 
 exports.parseConfigs = parseConfigs;
 exports.mergeChildConfig = mergeChildConfig;
 exports.filterConfig = filterConfig;
+exports.getManagerConfig = getManagerConfig;
+
+function getManagerConfig(config, manager) {
+  let managerConfig = config;
+  const language = get(manager, 'language');
+  if (language) {
+    managerConfig = mergeChildConfig(managerConfig, config[language]);
+  }
+  managerConfig = mergeChildConfig(managerConfig, config[manager]);
+  for (const i of getLanguageList().concat(getManagerList())) {
+    delete managerConfig[i];
+  }
+  return managerConfig;
+}
 
 async function parseConfigs(env, argv) {
   logger.debug('Parsing configs');
