+++ lib/workers/repository/process/extract-update.js
@@ -1,9 +1,7 @@
 const { writeUpdates } = require('./write');
 const { sortBranches } = require('./sort');
-const {
-  extractDependencies,
-  fetchUpdates,
-} = require('../../../../lib/manager');
+const { get } = require('../../../manager');
+const { fetchUpdates } = require('./fetch');
 const { branchifyUpgrades } = require('../updates/branchify');
 
 module.exports = {
@@ -11,6 +9,7 @@ module.exports = {
 };
 
 async function extractAndUpdate(config) {
+  const extractDependencies = get(config.manager, 'extractDependencies');
   const packageFiles = await extractDependencies(config);
   logger.debug({ packageFiles }, 'packageFiles');
   await fetchUpdates(config, packageFiles);
