+++ lib/workers/repository/updates/branchify.js
@@ -3,7 +3,7 @@ const slugify = require('slugify');
 const cleanGitRef = require('clean-git-ref').clean;
 
 const { generateBranchConfig } = require('./generate');
-const { flattenUpdates } = require('../../../manager');
+const { flattenUpdates } = require('./flatten');
 
 /**
  * Clean git branch name
